+++ jdto/src/main/java/org/jdto/mergers/AgeMerger.java
@@ -95,6 +95,17 @@ public class AgeMerger implements SinglePropertyValueMerger<Double, Object> {
         
     }
 
+    @Override
+    public boolean isUnmergeSupported(String[] params) {
+        //there is no way yo know the original date by just an age.
+        return false;
+    }
+
+    @Override
+    public Object unmergeObject(Double object, String[] params) {
+        return null; //to fulfill the contract.
+    }
+    
     private long getOtherStamp(Object value) {
         if (value instanceof Date) {
             Date ret = (Date) value;