+++ jdto/src/main/java/org/jdto/mergers/CloneMerger.java
@@ -15,8 +15,8 @@
  */
 package org.jdto.mergers;
 
-import org.jdto.SinglePropertyValueMerger;
 import java.lang.reflect.Method;
+import org.jdto.SinglePropertyValueMerger;
 
 /**
  * Merge the object by calling {@link java.lang.Object#clone() clone()}. <br />
@@ -58,5 +58,15 @@ public class CloneMerger implements SinglePropertyValueMerger<Object, Object> {
         } catch (Exception ex) {
             return null;
         }
+    }
+
+    @Override
+    public boolean isUnmergeSupported(String[] params) {
+        return false; //I cant return the orginal object from a clone, just another clone.
+    }
+
+    @Override
+    public Object unmergeObject(Object object, String[] params) {
+        return null;
     }
 }