+++ jdto/src/main/java/org/jdto/mergers/SumExpressionMerger.java
@@ -81,4 +81,14 @@ public class SumExpressionMerger extends AbstractCalulationCollectionMerger {
 
         return sum;
     }
+
+    @Override
+    public boolean isUnmergeSupported(String[] params) {
+        return false;
+    }
+
+    @Override
+    public Object unmergeObject(Double object, String[] params) {
+        return null;
+    }
 }