+++ jdto/src/main/java/org/jdto/mergers/SumMerger.java
@@ -56,4 +56,14 @@ public class SumMerger extends AbstractCalulationCollectionMerger {
 
         return sum;
     }
+
+    @Override
+    public boolean isUnmergeSupported(String[] params) {
+        return false;
+    }
+
+    @Override
+    public Object unmergeObject(Double object, String[] params) {
+        return null;
+    }
 }